update docs, move counter
int

int

int
