+++ docs/guide/counter/cmd/counter/main.go
@@ -5,8 +5,11 @@ import (
 
 	"github.com/spf13/cobra"
 
-	"github.com/tendermint/basecoin/cmd/commands"
 	"github.com/tendermint/tmlibs/cli"
+
+	"github.com/tendermint/basecoin/cmd/commands"
+	"github.com/tendermint/basecoin/docs/guide/counter/plugins/counter"
+	"github.com/tendermint/basecoin/types"
 )
 
 func main() {
@@ -22,6 +25,7 @@ func main() {
 		commands.VersionCmd,
 	)
 
-	cmd := cli.PrepareMainCmd(RootCmd, "BC", os.ExpandEnv("$HOME/.basecoin"))
+	commands.RegisterStartPlugin("counter", func() types.Plugin { return counter.New() })
+	cmd := cli.PrepareMainCmd(RootCmd, "CT", os.ExpandEnv("$HOME/.counter"))
 	cmd.Execute()
 }
