+++ docs/guide/counter/plugins/counter/counter.go
