+++ docs/guide/counter/plugins/counter/counter_test.go
