+++ docs/guide/install.md
@@ -3,7 +3,8 @@
 On a good day, basecoin can be installed like a normal Go program:
 
 ```
-go get -u github.com/tendermint/basecoin/cmd/basecoin
+go get github.com/tendermint/basecoin/
+make all
 ```
 
 In some cases, if that fails, or if another branch is required,
@@ -15,7 +16,7 @@ the correct way to install is:
 cd $GOPATH/src/github.com/tendermint/basecoin
 git pull origin master
 make get_vendor_deps
-make install
+make all
 ```
 
 This will create the `basecoin` binary in `$GOPATH/bin`.
